feat: refactor Options.tsx into components and update architecture accordingly
